Refactored classes to create an UpdateLineChart and UpdatePieChart class
